+++ bench-vortex/src/tpch_benchmark.rs
@@ -24,19 +24,19 @@ use crate::{BenchmarkDataset, Format, IdempotentPath, Target};
 
 /// TPCH benchmark implementation
 pub struct TpcHBenchmark {
-    pub scale_factor: u32,
+    pub scale_factor: f32,
     pub data_url: Url,
 }
 
 impl TpcHBenchmark {
-    pub fn new(scale_factor: u32, use_remote_data_dir: Option<String>) -> Result<Self> {
+    pub fn new(scale_factor: f32, use_remote_data_dir: Option<String>) -> Result<Self> {
         Ok(Self {
             scale_factor,
             data_url: Self::create_data_url(&use_remote_data_dir, scale_factor)?,
         })
     }
 
-    fn create_data_url(remote_data_dir: &Option<String>, scale_factor: u32) -> Result<Url> {
+    fn create_data_url(remote_data_dir: &Option<String>, scale_factor: f32) -> Result<Url> {
         match remote_data_dir {
             None => {
                 let data_dir = "tpch".to_data_path();
@@ -134,8 +134,8 @@ impl Benchmark for TpcHBenchmark {
 
     fn expected_row_counts(&self) -> Option<&[usize]> {
         match self.scale_factor {
-            1 => Some(&EXPECTED_ROW_COUNTS_SF1),
-            10 => Some(&EXPECTED_ROW_COUNTS_SF10),
+            1.0 => Some(&EXPECTED_ROW_COUNTS_SF1),
+            10.0 => Some(&EXPECTED_ROW_COUNTS_SF10),
             _ => None, // Unsupported scale factor
         }
     }
@@ -234,7 +234,7 @@ impl TpcHBenchmark {
     /// Verify DuckDB TPCH results against reference data
     pub fn verify_duckdb_tpch_results(&self, queries: Vec<usize>) -> Result<()> {
         // omit validation for sf != 1.
-        if self.scale_factor != 1 {
+        if self.scale_factor != 1.0 {
             return Ok(());
         }
         let query_dir = Path::new(env!("CARGO_MANIFEST_DIR"))
